Rename PiImage -> pigui.libs.image
