+++ data/pigui/modules/station-view/01-lobby.lua
@@ -14,7 +14,7 @@ local Comms = require 'Comms'
 local Space = require 'Space'
 
 local PiGuiFace = require 'pigui.libs.face'
-local PiImage = require 'ui.PiImage'
+local PiImage = require 'pigui.libs.image'
 local textTable = require 'pigui.libs.text-table'
 local ModalWindow = require 'pigui.libs.modal-win'
